+++ src/routes/NewsStory.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+function NewsStory() {
+  return <div>NewsStory</div>;
+}
+
+export default NewsStory;
